% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_cnv_depth.R
\name{binned_depth}
\alias{binned_depth}
\title{import copy number depth bedgraph and process to 5k-binned}
\usage{
binned_depth(file, keep_seq_levels = paste0("chr", 1:22), bin_size = 5000L)
}
\arguments{
\item{file}{A numeric value to be doubled.}
}
\value{
A data.frame of the copy number depth information from hificnv
}
\description{
This function takes a path of the depth's bedgraph
}
\examples{
my_function(5) # Returns 10
my_function(10) # Returns 20
}
