% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_cnv_vcf}
\alias{read_cnv_vcf}
\title{Import VCF (gz) to Granges object}
\usage{
read_cnv_vcf(file, genome = null, keep_seq_levels = NULL)
}
\arguments{
\item{file}{A path to the VCF (GZ) file}

\item{genome}{A character string of Genome identifier corresponding to chromosome names in the file. This identifier replaces the genome information in the VCF Seqinfo (i.e., \code{seqinfo(vcf)}). When not provided, genome is taken from the VCF file header}

\item{keep_seq_levels}{A character vector indicating which \code{seqlevels} to keep}
}
\value{
A \code{GRanges} object embedding variant information in the metadata columns
}
\description{
A wrapper \code{VariantAnnotation::readVcf} to tidy up the VCF and convert to GRanges
}
\seealso{
\code{\link[VariantAnnotation:readVcf-methods]{VariantAnnotation::readVcf()}}
}
