% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_copynumber_bedgraph.R
\name{read_copynumber_bedgraph}
\alias{read_copynumber_bedgraph}
\title{Import HiFiCNV output copynumber.bedgraph and convert to Granges}
\usage{
read_copynumber_bedgraph(file, keep_seq_levels)
}
\arguments{
\item{file}{A path to the VCF (GZ) file}

\item{keep_seq_levels}{A character vector indicating which seqlevels to keep}
}
\value{
A \code{GRanges} object with copy number variant type annotation
}
\description{
A wrapper of plyranges::read_bed_graph() to read HiFiCNV output copynumber.bedgraph file, convert to GRanges, and annotation the copy number score
}
\details{
Copy number variant type (SVTYPE) = "DUP" if score > 2 and = "DEL" if score = 0. NA otherwise.
}
