% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ft_extract_all.R
\name{read_ft_extract_all}
\alias{read_ft_extract_all}
\title{Title: import ft extract <input.bam> --all bed.gz file}
\usage{
read_ft_extract_all(
  file,
  col_names = NULL,
  genome_info = NULL,
  overlap_ranges = NULL
)
}
\arguments{
\item{file}{A path to the BED GZ file yielded by ft extract --all}

\item{col_names}{An optional characger vector for including additional columns in file that is not part of the BED format}

\item{genome_inf}{An optional character or Ranges object that contains information about the genome (e.g., 'hg38')}

\item{overlap_ragnes}{An optional \code{GRanges} in the file that overlap the ranges will be return.}
}
\value{
A \code{GRanges} object with 26 additional columns specified in the fibertools-rs documentation.
}
\description{
Tool to import ft extract --all output file
}
\references{
The computational guide to Fiber-seq \url{https://fiberseq.github.io/fibertools/extracting/extract.html}
}
