% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Import VCF (gz) to Granges object}
\usage{
read_vcf(file, keep_seq_levels = NULL)
}
\arguments{
\item{file}{A path to the VCF (GZ) file}

\item{keep_seq_levels}{A character vector indicating which seqlevels to keep}
}
\value{
A \code{GRanges} object embedding variant information in the metadata columns
}
\description{
A wrapper VCF import function of VariantAnnotation to tidy up the VCF and convert to GRanges
}
